+++ scripts/lichessTools-init.js
@@ -144,6 +144,7 @@ lichessTools.loadTool(LiChessTools.Tools.ChessPursuit);
 lichessTools.loadTool(LiChessTools.Tools.LobbyRatingDiff);
 lichessTools.loadTool(LiChessTools.Tools.VoiceInput);
 lichessTools.loadTool(LiChessTools.Tools.BotsMenuItem);
+lichessTools.loadTool(LiChessTools.Tools.TeamStats);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);